+++ test/internal-config-specs.js
@@ -108,22 +108,29 @@ describe("When a new t-motion-detector instance is started from main, ", functio
   });
 
   it('the object Config can accept alternative mock config files', function (done) {
-    let alternativeConfig = new main.Config("./test/config_test1.js");
+    let alternativeConfig = new main.Config("/test/config_test1.js");
     let file = require("./config_test1.js");
     alternativeConfig.file.should.equal(file);
     done();
   });
 
   it('the object Config can detect the default profile of a mock config file', function (done) {
-    let alternativeConfig = new main.Config("./test/config_test1.js");
+    let alternativeConfig = new main.Config("/test/config_test1.js");
     let file_val = alternativeConfig.profile("default");
 
     JSON.stringify(file_val.some_group).should.equal(JSON.stringify({some_property: "Test Property" }));
     done();
   });
 
+  it('if the object Config cannot detect the Configuration file, it fallsback to default and highlights through the isFallBack() function', function (done) {
+    let alternativeConfig = new main.Config("/test/config_test_does_not_exist.js");
+
+    alternativeConfig.isFallback().should.equal(true);
+    done();
+  });
+
   it('the object Config can detect properties from mock config files', function (done) {
-    let alternativeConfig = new main.Config("./test/config_test1.js");
+    let alternativeConfig = new main.Config("/test/config_test1.js");
 
     alternativeConfig.getProperty("default", "some_group").some_property.should.equal("Test Property");
     done();
@@ -177,7 +184,7 @@ describe("When a new t-motion-detector instance is started from main, ", functio
   });
 
   it('the object Config can receive profiles with the name of the class (dependency injection), and array of attributes', function (done) {
-    let alternativeConfig = new main.Config("./test/config_test2.js");
+    let alternativeConfig = new main.Config("/test/config_test2.js");
     main.StartWithConfig(alternativeConfig);
 
     let slackNotifier = main.GetNotifiers()[0];
@@ -189,7 +196,7 @@ describe("When a new t-motion-detector instance is started from main, ", functio
   it('When starting with alternate config, the Notification Objects contained in the config file are automatically instanced as Notificators of main', function (done) {
     //Main needs to be reset explicitely because it keeps objects from previous test
     main.Reset();
-    let alternativeConfig = new main.Config("./test/config_test2.js");
+    let alternativeConfig = new main.Config("/test/config_test2.js");
     main.StartWithConfig(alternativeConfig);
     main.GetNotifiers().length.should.equal(1);
 
@@ -199,7 +206,7 @@ describe("When a new t-motion-detector instance is started from main, ", functio
   it('When starting with alternate config, the Environment Objects contained in the config file are automatically instanced as Environments of main', function (done) {
     //Main needs to be reset explicitely because it keeps objects from previous test
     main.Reset();
-    let alternativeConfig = new main.Config("./test/config_test2.js");
+    let alternativeConfig = new main.Config("/test/config_test2.js");
     main.StartWithConfig(alternativeConfig);
 
     main.GetNotifiers().length.should.equal(1);
@@ -208,7 +215,7 @@ describe("When a new t-motion-detector instance is started from main, ", functio
   });
 
   it('When starting with alternate config, the MotionDetector Objects contained in the config file are automatically instanced as MotionDetectors of main', function (done) {
-    let alternativeConfig = new main.Config("./test/config_test2.js");
+    let alternativeConfig = new main.Config("/test/config_test2.js");
     main.StartWithConfig(alternativeConfig);
 
     let slackNotifier = main.GetNotifiers()[0];